Apply patch so fluxcdv1 friendly
