+++ charts/istio/1.18.2/istio-control/istio-discovery/templates/revision-tags.yaml
@@ -1,11 +1,7 @@
 # Adapted from istio-discovery/templates/mutatingwebhook.yaml
 # Removed paths for legacy and default selectors since a revision tag
 # is inherently created from a specific revision
-{{- $whv := dict
- "revision" .Values.revision
-  "injectionPath" .Values.istiodRemote.injectionPath
-  "injectionURL" .Values.istiodRemote.injectionURL
-  "namespace" .Release.Namespace }}
+{{- $whv := dict "revision" .Values.revision "injectionPath" .Values.istiodRemote.injectionPath "injectionURL" .Values.istiodRemote.injectionURL "namespace" .Release.Namespace }}
 {{- define "core" }}
 {{- /* Kubernetes unfortunately requires a unique name for the webhook in some newer versions, so we assign
 a unique prefix to each. */}}
