+++ config.yaml
@@ -1,20 +1,20 @@
 istios:
-  - version: "1.13.3"
-  - version: "1.13.5"
-  - version: "1.14.1"
-  - version: "1.14.3"
-  - version: "1.14.5"
-  - version: "1.14.6"
-  - version: "1.15.4"
-  - version: "1.16.2"
-  - version: "1.16.4"
-    tetrate: "4"
-  - version: "1.16.5"
-  - version: "1.16.7"
-  - version: "1.17.1"
-  - version: "1.17.3"
-  - version: "1.17.5"
-  - version: "1.18.0"
+  # - version: "1.13.3"
+  # - version: "1.13.5"
+  # - version: "1.14.1"
+  # - version: "1.14.3"
+  # - version: "1.14.5"
+  # - version: "1.14.6"
+  # - version: "1.15.4"
+  # - version: "1.16.2"
+  # - version: "1.16.4"
+  #   tetrate: "4"
+  # - version: "1.16.5"
+  # - version: "1.16.7"
+  # - version: "1.17.1"
+  # - version: "1.17.3"
+  # - version: "1.17.5"
+  # - version: "1.18.0"
   - version: "1.18.2"
 
 addons:
